Sanity check + cleanup
Added a sanity check to the manual value and removed a annoying warning

+++ OpenVR-SpaceCalibrator/Calibration.cpp
@@ -444,7 +444,7 @@ void CalibrationTick(double time)
 		ctx.state = CalibrationState::Rotation;
 		ctx.wantedUpdateInterval = 0.0;
 
-		snprintf(buf, sizeof buf, "Starting calibration with %d samples...\n", CalCtx.SampleCount());
+		snprintf(buf, sizeof buf, "Starting calibration with %d samples...\n", (int) CalCtx.SampleCount());
 		CalCtx.Log(buf);
 		return;
 	}
